spotless

+++ src/main/java/com/ytdd9527/networksexpansion/utils/ParticleUtil.java
@@ -1,15 +1,14 @@
 package com.ytdd9527.networksexpansion.utils;
 
 import io.github.sefiraat.networks.Networks;
+import java.util.ArrayList;
+import java.util.List;
+import javax.annotation.Nonnull;
 import org.bukkit.Location;
 import org.bukkit.Particle;
 import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.plugin.Plugin;
-
-import javax.annotation.Nonnull;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author Final_ROOT
@@ -17,11 +16,12 @@ import java.util.List;
  * @since 2.0
  */
 public class ParticleUtil {
-    private static final double[] BLOCK_CUBE_OFFSET_X = new double[]{0, 1, 0, 0, 1, 1, 0, 1};
-    private static final double[] BLOCK_CUBE_OFFSET_Y = new double[]{0, 0, 1, 0, 1, 0, 1, 1};
-    private static final double[] BLOCK_CUBE_OFFSET_Z = new double[]{0, 0, 0, 1, 0, 1, 1, 1};
+    private static final double[] BLOCK_CUBE_OFFSET_X = new double[] {0, 1, 0, 0, 1, 1, 0, 1};
+    private static final double[] BLOCK_CUBE_OFFSET_Y = new double[] {0, 0, 1, 0, 1, 0, 1, 1};
+    private static final double[] BLOCK_CUBE_OFFSET_Z = new double[] {0, 0, 0, 1, 0, 1, 1, 1};
 
-    public static void drawLineByTotalAmount(@Nonnull Particle particle, int totalAmount, @Nonnull Location... locations) {
+    public static void drawLineByTotalAmount(
+            @Nonnull Particle particle, int totalAmount, @Nonnull Location... locations) {
         for (int i = 0; i < locations.length; i++) {
             if ((i + 1) < locations.length) {
                 Location location1 = locations[i];
@@ -41,7 +41,8 @@ public class ParticleUtil {
         }
     }
 
-    public static void drawLineByTotalAmount(@Nonnull Particle particle, int totalAmount, @Nonnull List<Location> locationList) {
+    public static void drawLineByTotalAmount(
+            @Nonnull Particle particle, int totalAmount, @Nonnull List<Location> locationList) {
         Location[] locations = new Location[locationList.size()];
         for (int i = 0; i < locations.length; i++) {
             locations[i] = locationList.get(i);
@@ -49,7 +50,12 @@ public class ParticleUtil {
         ParticleUtil.drawLineByTotalAmount(particle, totalAmount, locations);
     }
 
-    public static void drawLineByDistance(@Nonnull Plugin plugin, @Nonnull Particle particle, long interval, double distance, @Nonnull Location... locations) {
+    public static void drawLineByDistance(
+            @Nonnull Plugin plugin,
+            @Nonnull Particle particle,
+            long interval,
+            double distance,
+            @Nonnull Location... locations) {
         int time = 0;
         for (int i = 0; i + 1 < locations.length; i++) {
             Location location1 = locations[i];
@@ -86,21 +92,30 @@ public class ParticleUtil {
                 lastTick = (int) (t / 50);
                 if (tick != lastTick) {
                     final List<Runnable> finalRunnableList = runnableList;
-                    plugin.getServer().getScheduler().runTaskLaterAsynchronously(plugin, () -> finalRunnableList.forEach(Runnable::run), tick);
+                    plugin.getServer()
+                            .getScheduler()
+                            .runTaskLaterAsynchronously(plugin, () -> finalRunnableList.forEach(Runnable::run), tick);
                     tick = lastTick;
                     runnableList = new ArrayList<>();
                 }
             }
             if (!runnableList.isEmpty()) {
                 final List<Runnable> finalRunnableList = runnableList;
-                plugin.getServer().getScheduler().runTaskLaterAsynchronously(plugin, () -> finalRunnableList.forEach(Runnable::run), tick);
+                plugin.getServer()
+                        .getScheduler()
+                        .runTaskLaterAsynchronously(plugin, () -> finalRunnableList.forEach(Runnable::run), tick);
             }
 
             time += (int) interval;
         }
     }
 
-    public static void drawLineByDistance(@Nonnull Plugin plugin, @Nonnull Particle particle, long interval, double distance, @Nonnull List<Location> locationList) {
+    public static void drawLineByDistance(
+            @Nonnull Plugin plugin,
+            @Nonnull Particle particle,
+            long interval,
+            double distance,
+            @Nonnull List<Location> locationList) {
         Location[] locations = new Location[locationList.size()];
         for (int i = 0; i < locations.length; i++) {
             locations[i] = locationList.get(i);
@@ -108,7 +123,8 @@ public class ParticleUtil {
         ParticleUtil.drawLineByDistance(plugin, particle, interval, distance, locations);
     }
 
-    public static void drawCubeByLocations(@Nonnull Plugin plugin, @Nonnull Particle particle, long interval, Location... locations) {
+    public static void drawCubeByLocations(
+            @Nonnull Plugin plugin, @Nonnull Particle particle, long interval, Location... locations) {
         int time = 0;
         for (Location location : locations) {
             World world = location.getWorld();
@@ -120,20 +136,44 @@ public class ParticleUtil {
             int z = location.getBlockZ();
             if (time < 50) {
                 for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
-                    world.spawnParticle(particle, x + BLOCK_CUBE_OFFSET_X[i], y + BLOCK_CUBE_OFFSET_Y[i], z + BLOCK_CUBE_OFFSET_Z[i], 1, 0, 0, 0, 0);
+                    world.spawnParticle(
+                            particle,
+                            x + BLOCK_CUBE_OFFSET_X[i],
+                            y + BLOCK_CUBE_OFFSET_Y[i],
+                            z + BLOCK_CUBE_OFFSET_Z[i],
+                            1,
+                            0,
+                            0,
+                            0,
+                            0);
                 }
             } else {
-                plugin.getServer().getScheduler().runTaskLaterAsynchronously(plugin, () -> {
-                    for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
-                        world.spawnParticle(particle, x + BLOCK_CUBE_OFFSET_X[i], y + BLOCK_CUBE_OFFSET_Y[i], z + BLOCK_CUBE_OFFSET_Z[i], 1, 0, 0, 0, 0);
-                    }
-                }, time / 50);
+                plugin.getServer()
+                        .getScheduler()
+                        .runTaskLaterAsynchronously(
+                                plugin,
+                                () -> {
+                                    for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
+                                        world.spawnParticle(
+                                                particle,
+                                                x + BLOCK_CUBE_OFFSET_X[i],
+                                                y + BLOCK_CUBE_OFFSET_Y[i],
+                                                z + BLOCK_CUBE_OFFSET_Z[i],
+                                                1,
+                                                0,
+                                                0,
+                                                0,
+                                                0);
+                                    }
+                                },
+                                time / 50);
             }
             time += (int) interval;
         }
     }
 
-    public static void drawCubeByLocations(@Nonnull Plugin plugin, @Nonnull Particle particle, long interval, List<Location> locationList) {
+    public static void drawCubeByLocations(
+            @Nonnull Plugin plugin, @Nonnull Particle particle, long interval, List<Location> locationList) {
         Location[] locations = new Location[locationList.size()];
         for (int i = 0; i < locationList.size(); i++) {
             locations[i] = locationList.get(i);
@@ -141,7 +181,12 @@ public class ParticleUtil {
         ParticleUtil.drawCubeByLocations(plugin, particle, interval, locations);
     }
 
-    public static void drawRegionOutline(@Nonnull Plugin plugin, @Nonnull Particle particle, long interval, @Nonnull Location corner1, @Nonnull Location corner2) {
+    public static void drawRegionOutline(
+            @Nonnull Plugin plugin,
+            @Nonnull Particle particle,
+            long interval,
+            @Nonnull Location corner1,
+            @Nonnull Location corner2) {
         World world = corner1.getWorld();
         if (world == null || corner1.getWorld() != corner2.getWorld()) {
             return;
@@ -154,15 +199,15 @@ public class ParticleUtil {
         double maxY = Math.max(corner1.getY(), corner2.getY());
         double maxZ = Math.max(corner1.getZ(), corner2.getZ());
 
-        Location[] corners = new Location[]{
-                new Location(world, minX, minY, minZ),
-                new Location(world, minX, minY, maxZ + 1),
-                new Location(world, maxX + 1, minY, maxZ + 1),
-                new Location(world, maxX + 1, minY, minZ),
-                new Location(world, minX, maxY + 1, minZ),
-                new Location(world, minX, maxY + 1, maxZ + 1),
-                new Location(world, maxX + 1, maxY + 1, maxZ + 1),
-                new Location(world, maxX + 1, maxY + 1, minZ)
+        Location[] corners = new Location[] {
+            new Location(world, minX, minY, minZ),
+            new Location(world, minX, minY, maxZ + 1),
+            new Location(world, maxX + 1, minY, maxZ + 1),
+            new Location(world, maxX + 1, minY, minZ),
+            new Location(world, minX, maxY + 1, minZ),
+            new Location(world, minX, maxY + 1, maxZ + 1),
+            new Location(world, maxX + 1, maxY + 1, maxZ + 1),
+            new Location(world, maxX + 1, maxY + 1, minZ)
         };
 
         drawLineByDistance(plugin, particle, interval, 0.25, corners[0], corners[1]);
@@ -184,7 +229,11 @@ public class ParticleUtil {
         drawLineByDistance(plugin, particle, interval, 0.25, corners[6], corners[7]);
     }
 
-    public static void drawCubeByBlock(@Nonnull final Plugin plugin, @Nonnull final Particle particle, final long interval, @Nonnull final Block... blocks) {
+    public static void drawCubeByBlock(
+            @Nonnull final Plugin plugin,
+            @Nonnull final Particle particle,
+            final long interval,
+            @Nonnull final Block... blocks) {
         int time = 0;
         for (final Block block : blocks) {
             final Location location = block.getLocation();
@@ -197,20 +246,47 @@ public class ParticleUtil {
             final int z = location.getBlockZ();
             if (time < 50) {
                 for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
-                    world.spawnParticle(particle, x + BLOCK_CUBE_OFFSET_X[i], y + BLOCK_CUBE_OFFSET_Y[i], z + BLOCK_CUBE_OFFSET_Z[i], 1, 0, 0, 0, 0);
+                    world.spawnParticle(
+                            particle,
+                            x + BLOCK_CUBE_OFFSET_X[i],
+                            y + BLOCK_CUBE_OFFSET_Y[i],
+                            z + BLOCK_CUBE_OFFSET_Z[i],
+                            1,
+                            0,
+                            0,
+                            0,
+                            0);
                 }
             } else {
-                plugin.getServer().getScheduler().runTaskLaterAsynchronously(plugin, () -> {
-                    for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
-                        world.spawnParticle(particle, x + BLOCK_CUBE_OFFSET_X[i], y + BLOCK_CUBE_OFFSET_Y[i], z + BLOCK_CUBE_OFFSET_Z[i], 1, 0, 0, 0, 0);
-                    }
-                }, time / 50);
+                plugin.getServer()
+                        .getScheduler()
+                        .runTaskLaterAsynchronously(
+                                plugin,
+                                () -> {
+                                    for (int i = 0; i < BLOCK_CUBE_OFFSET_X.length; i++) {
+                                        world.spawnParticle(
+                                                particle,
+                                                x + BLOCK_CUBE_OFFSET_X[i],
+                                                y + BLOCK_CUBE_OFFSET_Y[i],
+                                                z + BLOCK_CUBE_OFFSET_Z[i],
+                                                1,
+                                                0,
+                                                0,
+                                                0,
+                                                0);
+                                    }
+                                },
+                                time / 50);
             }
-            time += interval;
+            time += (int) interval;
         }
     }
 
-    public static void drawCubeByBlock(@Nonnull final Plugin plugin, @Nonnull final Particle particle, final long interval, final List<Block> blockList) {
+    public static void drawCubeByBlock(
+            @Nonnull final Plugin plugin,
+            @Nonnull final Particle particle,
+            final long interval,
+            final List<Block> blockList) {
         Block[] blocks = new Block[blockList.size()];
         for (int i = 0; i < blockList.size(); i++) {
             blocks[i] = blockList.get(i);
